+++ lib/src/gestures_listener.dart
@@ -6,7 +6,7 @@ import 'package:flutter_tilt/src/state/tilt_state.dart';
 import 'package:flutter_tilt/src/type/tilt_type.dart';
 
 /// 手势监听
-class GesturesListener extends StatelessWidget {
+class GesturesListener extends StatefulWidget {
   /// 手势监听
   ///
   /// 对 Touch [Listener] 和 Mouse [MouseRegion] 的监听触发
@@ -20,9 +20,20 @@ class GesturesListener extends StatelessWidget {
   final TiltConfig tiltConfig;
 
   @override
-  Widget build(BuildContext context) {
-    final TiltState tiltState = TiltState.of(context)!;
+  State<GesturesListener> createState() => _GesturesListenerState();
+}
 
+class _GesturesListenerState extends State<GesturesListener> {
+  Widget get _child => widget.child;
+  TiltConfig get _tiltConfig => widget.tiltConfig;
+
+  late TiltState tiltState = TiltState.of(context)!;
+
+  /// 避免 touch 和 hover 同时触发，导致 hover 离开的时候会“闪现”
+  bool isTouch = false;
+
+  @override
+  Widget build(BuildContext context) {
     /// 不受滑动影响
     return GestureDetector(
       onVerticalDragUpdate: (_) {},
@@ -30,28 +41,28 @@ class GesturesListener extends StatelessWidget {
 
       /// 手势监听
       child: Listener(
-        onPointerMove: (PointerMoveEvent e) => tiltState.onMove(
-          e.localPosition,
-          GesturesType.touch,
-        ),
-        onPointerUp: (PointerUpEvent e) => tiltState.onRevert(
-          e.localPosition,
-          GesturesType.touch,
-        ),
-        onPointerCancel: (PointerCancelEvent e) => tiltState.onRevert(
-          e.localPosition,
-          GesturesType.touch,
-        ),
+        onPointerMove: (PointerMoveEvent e) {
+          isTouch = true;
+          tiltState.onMove(e.localPosition, GesturesType.touch);
+        },
+        onPointerUp: (PointerUpEvent e) {
+          isTouch = false;
+          tiltState.onRevert(e.localPosition, GesturesType.touch);
+        },
+        onPointerCancel: (PointerCancelEvent e) {
+          isTouch = false;
+          tiltState.onRevert(e.localPosition, GesturesType.touch);
+        },
         child: MouseRegion(
-          onHover: tiltConfig.enableMouseHover
+          onHover: _tiltConfig.enableMouseHover && !isTouch
               ? (PointerHoverEvent e) =>
                   tiltState.onMove(e.localPosition, GesturesType.hover)
               : null,
-          onExit: tiltConfig.enableMouseHover
+          onExit: _tiltConfig.enableMouseHover && !isTouch
               ? (PointerExitEvent e) =>
                   tiltState.onRevert(e.localPosition, GesturesType.hover)
               : null,
-          child: child,
+          child: _child,
         ),
       ),
     );